receive screen

+++ app/src/main/java/com/cerberus/galtwallet/infrastructure/BDKWallet.kt
@@ -41,7 +41,7 @@ class BDKWallet @Inject constructor(
     }
 
     override fun getBalance(): Amount {
-        return Amount(wallet.getBalance())
+        return Amount(wallet.getBalance().toLong())
     }
 
     override fun getNewAddress(): Address {
